feat: Refactor check-in reset functionality (JasonLovesDoggo/add-checkbox-reset-endpoint)
The check-in reset functionality has been refactored to
improve efficiency and user experience.  The endpoint
has been moved to a DELETE request for better
semantic clarity.  The UI button text has been updated
for improved clarity.  Added a check to prevent
duplicate scans of the same user within a short time
frame.  Added `allowReset` config option to enable/disable
reset functionality.  Removed the dedicated reset
endpoint.
