+++ hooks/useQRScanner.ts
@@ -11,12 +11,16 @@ interface UseQRScannerProps {
   keepCameraOn: boolean;
 }
 
-export const useQRScanner = ({ selectedEvent, keepCameraOn }: UseQRScannerProps) => {
+export const useQRScanner = ({
+  selectedEvent,
+  keepCameraOn,
+}: UseQRScannerProps) => {
   const [isCameraOn, setIsCameraOn] = useState(false);
   const [startingCamera, setStartingCamera] = useState(false);
   const [scanResult, setScanResult] = useState<"success" | "error" | null>(
     null,
   );
+  const [lastUserId, setLastUserId] = useState<string | null>(null);
   const videoRef = useRef<HTMLVideoElement | null>(null);
   const codeReader = useRef(new BrowserMultiFormatReader());
   const isProcessing = useRef(false);
@@ -87,15 +91,19 @@ export const useQRScanner = ({ selectedEvent, keepCameraOn }: UseQRScannerProps)
     }
   };
 
-  const handleResetEvent = async (userId: string) => {
+  const handleResetEvent = async () => {
+    if (!lastUserId) {
+      toast.error("No user to reset, please scan a QR code first");
+      return;
+    }
     try {
-      const response = await fetch("/api/check-ins/reset", {
-        method: "POST",
+      const response = await fetch("/api/check-ins", {
+        method: "DELETE",
         headers: {
           "Content-Type": "application/json",
         },
         body: JSON.stringify({
-          userId,
+          userId: lastUserId,
           eventName: selectedEvent,
         }),
       });
@@ -126,6 +134,9 @@ export const useQRScanner = ({ selectedEvent, keepCameraOn }: UseQRScannerProps)
       return;
     }
 
+    let lastScan = new Date("1970-01-01");
+    let lastUserIdVar: String = ""; // to prevent double scans of the same user, we need this as react wont re-render in time to prevent double scans
+
     try {
       // Start the QR code reader directly with constraints
       await codeReader.current.decodeFromVideoDevice(
@@ -147,6 +158,20 @@ export const useQRScanner = ({ selectedEvent, keepCameraOn }: UseQRScannerProps)
             setTimeout(() => setScanResult(null), 1000);
             return;
           }
+
+          // prevent double scans of the same user
+          if (
+            (lastUserId === userId || lastUserIdVar == userId) &&
+            new Date().getTime() - lastScan.getTime() < 3000
+          ) {
+            // 3 seconds between scans of the same user to prevent double scans
+            // console.table({ lastUserIdVar, userId, lastScan: lastScan.getTime(), now: new Date().getTime() });
+            return;
+          }
+
+          lastScan = new Date();
+          lastUserIdVar = userId; // Due to the lack of rendering, we need to set this here to prevent double scans
+          setLastUserId(userId);
 
           isProcessing.current = true;
           await handleCheckIn(userId);
